Moves pg_uuid.py into app.utils.
